Fix pagination of timeline events
Before this commit, the timeline used indices to keep track of selected events. The problem with this is that the indices mean different things as the array of events is mutated based on filters. In other words, filtering the events lead to the indices pointing to different events.

Now the timeline uses IDs to keep track of selected events.